Update from llvm::main, change taint hooks between dfsan/polytracker

+++ compiler-rt/include/sanitizer/dfsan_interface.h
@@ -21,7 +21,7 @@
 extern "C" {
 #endif
 
-typedef uint16_t dfsan_label;
+typedef uint32_t dfsan_label;
 typedef uint32_t dfsan_origin;
 
 /// Stores information associated with a specific label identifier.  A label
